moved project to .NET 5, code refactoring

+++ ITCenter_dokumenty_magazynowe/Models/DbModels/WarehouseDoc.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
+
 
 namespace ITCenter_dokumenty_magazynowe.Models.DbModels
 {
@@ -18,9 +16,9 @@ namespace ITCenter_dokumenty_magazynowe.Models.DbModels
         [Required]
         public DateTime? Date { get; set; }
         [Required]
-        public int NetPrice { get; set; }
+        public double NetPrice { get; set; }
         [Required]
-        public int GrossPrice { get; set; }
+        public double GrossPrice { get; set; }
 
     }
 }
